+++ src/app/cart/cart.module.ts
@@ -6,13 +6,13 @@ import {CartItemComponent} from './cart-item/cart-item.component';
 import {CartService} from './cart.service';
 import {CartListComponent} from "./cart-list/cart-list.component";
 import {HeaderModule} from "../header/header.module";
-import {ChangeBackgroundDirective} from "./change-background.directive";
+import {CoreModule} from "../core/core.module";
 
 @NgModule({
   imports: [
-    CommonModule, FormsModule, HeaderModule
+    CommonModule, FormsModule, HeaderModule, CoreModule
   ],
-  declarations: [CartListComponent, CartItemComponent, ChangeBackgroundDirective],
+  declarations: [CartListComponent, CartItemComponent],
   providers: [CartService],
   exports: [CartListComponent]
 })