+++ src/app/core/service/config-options.service.ts
@@ -1,0 +1,11 @@
+import {Injectable} from '@angular/core';
+
+@Injectable()
+export class ConfigOptionsService {
+
+  constructor(private id: number,
+              private login: string,
+              private email: string) {
+  }
+
+}