+++ src/app/core/service/constants.service.ts
@@ -1,0 +1,10 @@
+import {Injectable} from '@angular/core';
+
+@Injectable()
+export class ConstantsService {
+
+  constructor(public Application: string,
+              public Version: string) {
+  }
+
+}