+++ src/app/header/header.module.ts
@@ -1,6 +1,5 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import {ProductListComponent} from "../product/product-list/product-list.component";
 import {HeaderComponent} from "./header.component";
 
 @NgModule({