+++ src/app/product/model/product.ts
@@ -1,11 +1,11 @@
 import {Category} from "./category";
 
 export interface Product {
-  id: number,
-  name: string,
-  description: string,
-  price: number,
-  category: Category,
-  isAvailable: boolean,
+  id: number;
+  name: string;
+  description: string;
+  price: number;
+  category: Category;
+  isAvailable: boolean;
   stores?: string[]
 }