+++ src/app/product/product-item/product-item.component.ts
@@ -1,17 +1,24 @@
 import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
 import {Product} from "../model/product";
+import {CHARACTERS_09, GeneratorService, GeneratorServiceFactory} from "../../core/service/generator.service";
+
 
 @Component({
   selector: 'product-item',
   templateUrl: './product-item.component.html',
-  styleUrls: ['./product-item.component.css']
+  styleUrls: ['./product-item.component.css'],
+  providers: [
+      {provide: GeneratorService, useFactory: GeneratorServiceFactory(10, CHARACTERS_09)}
+      ]
 })
 export class ProductItemComponent implements OnInit {
 
   @Input() product: Product;
   @Output() addToCart: EventEmitter<Product> = new EventEmitter<Product>();
+  hash: string;
 
-  constructor() {
+  constructor(private generatorService: GeneratorService) {
+    this.hash = this.generatorService.generate();
   }
 
   ngOnInit() {